% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{FizzBuzz function!}
\usage{
FizzBuzz(x)
}
\arguments{
\item{x}{A vector of integers.}
}
\description{
This is a function that shows output as a result from a childrens game with
numbers and their factors.
}
