% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{A time series of daily maximum temperatures in Flagstaff, AZ.}
\format{
Weather information at Flagstaff's Pulliam Airport with 22967 obs.
\describe{
  \item{DATE}{The date of observation as a POSIX date format.}
  \item{PRCP}{Precipitation.}
  \item{snowfall}{Snowfall.}
  \item{TMAX}{Maximum temperature.}
  \item{TMIN}{Minimum temperature.}
}
}
\source{
\url{www.ncdc.noaa.gov}
}
\usage{
Flagstaff_Weather
}
\description{
A time series of daily maximum temperatures in Flagstaff, AZ.
}
\keyword{datasets}
